Run karma in the process-test-classes phase by default.
This is the last phase before test and allows karma to be run BEFORE maven-surefire-plugin in a war project.

That way, test results of karma (using a junit-reporter) can automatically be picked up by CI servers listening explicitly for the result of surefire (Jenkins and Hudson).

Otherwise, karma would run after surefire, and when Jenkins reads test result, it would not get karma result.