Clarify README to state download roots must point to the same repo
Regarding download roots, the install instructions state that node and npm may have different roots "as they are now stored in separate repos".  However, the example implies the repo must be the same for both roots.
This change rewords that sentence to better convey that, while the download roots may be different, the downloads must come from the same repo.