Support environment file next to environment variables
With these changes the support is provided for environment files (eg. ".env") - basically any key-value pair based files.

If provided, their content is loaded and if used alongside the already existing 'environmentVariables' configuration, then they are merged.

Integration tests are provided to the already existing behavior as well as the new one and a mix of both.