Support corepack.
This adds some basic support for using corepack to manage the tooling
via the 'packageManager' field in the package.json. Here we add the
ability to download a version of it along with a node version, along
with the ability to execute command via corepack such as 'pnpm install'
or 'yarn install' depending on the selected package manager.

Later we ought to work out how to use the version packaged with the node
runtime, but this was inspired by the pnpm download which is downloaded
separately, and so this is too.