DCA11Y-1315: Use PAC mirrors where possible
Excuse the factoring, it's got enough spaghetti for a whole 'nother generation of Eminem memes. Why? Trying to make it easy to update from upstream. If I was to refactor the existing code we're more likely to get merge-conflicts. I don't see us ever being able to upstream this change so it doesn't make sense for the diff to be optimised for that.